feat: move animation files to src/animations/
