+++ src/animations/alt_chest_opening.c
