+++ src/animations/alt_chest_opening.h
@@ -1,7 +1,7 @@
 #ifndef ALT_CHEST_OPENING_S4C_H_
 #define ALT_CHEST_OPENING_S4C_H_
 #define ALT_CHEST_OPENING_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for alt_chest_opening.
