+++ src/animations/archer_drop.h
@@ -1,7 +1,7 @@
 #ifndef ARCHER_DROP_S4C_H_
 #define ARCHER_DROP_S4C_H_
 #define ARCHER_DROP_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for archer_drop.
