+++ src/animations/assassin_poof.c
