+++ src/animations/assassin_poof.h
@@ -1,7 +1,7 @@
 #ifndef ASSASSIN_POOF_S4C_H_
 #define ASSASSIN_POOF_S4C_H_
 #define ASSASSIN_POOF_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for assassin_poof.
