+++ src/animations/bluetroll_wonder.h
@@ -1,7 +1,7 @@
 #ifndef BLUETROLL_WONDER_S4C_H_
 #define BLUETROLL_WONDER_S4C_H_
 #define BLUETROLL_WONDER_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for bluetroll_wonder.
