+++ src/animations/boar_scream.c
