+++ src/animations/boar_scream.h
@@ -1,7 +1,7 @@
 #ifndef BOAR_SCREAM_S4C_H_
 #define BOAR_SCREAM_S4C_H_
 #define BOAR_SCREAM_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for boar_scream.
