+++ src/animations/crawlingdude_crawl.c
