+++ src/animations/crawlingdude_crawl.h
@@ -1,7 +1,7 @@
 #ifndef CRAWLINGDUDE_CRAWL_S4C_H_
 #define CRAWLINGDUDE_CRAWL_S4C_H_
 #define CRAWLINGDUDE_CRAWL_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for crawlingdude_crawl.
