+++ src/animations/enter_door.c
