+++ src/animations/enter_door.h
@@ -1,7 +1,7 @@
 #ifndef ENTER_DOOR_S4C_H_
 #define ENTER_DOOR_S4C_H_
 #define ENTER_DOOR_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for enter_door.
