+++ src/animations/ghost_spell.c
