+++ src/animations/ghost_spell.h
@@ -1,7 +1,7 @@
 #ifndef GHOST_SPELL_S4C_H_
 #define GHOST_SPELL_S4C_H_
 #define GHOST_SPELL_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for ghost_spell.
