+++ src/animations/goblin_shoot.c
