+++ src/animations/goblin_shoot.h
@@ -1,7 +1,7 @@
 #ifndef GOBLIN_SHOOT_S4C_H_
 #define GOBLIN_SHOOT_S4C_H_
 #define GOBLIN_SHOOT_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for goblin_shoot.
