+++ src/animations/headlessninja_throw.h
@@ -1,6 +1,7 @@
 #ifndef HEADLESSNINJA_THROW_S4C_H_
+#define HEADLESSNINJA_THROW_S4C_H_
 #define HEADLESSNINJA_THROW_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for headlessninja_throw.
