+++ src/animations/imp_fireball.h
@@ -1,7 +1,7 @@
 #ifndef IMP_FIREBALL_S4C_H_
 #define IMP_FIREBALL_S4C_H_
 #define IMP_FIREBALL_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for imp_fireball.
