+++ src/animations/knight_tapis.c
