+++ src/animations/knight_tapis.h
@@ -1,6 +1,7 @@
 #ifndef KNIGHT_TAPIS_S4C_H_
+#define KNIGHT_TAPIS_S4C_H_
 #define KNIGHT_TAPIS_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for knight_tapis.
