+++ src/animations/mage_spark.c
