+++ src/animations/mummy_shuffle.c
