+++ src/animations/mummy_shuffle.h
@@ -1,7 +1,7 @@
 #ifndef MUMMY_SHUFFLE_S4C_H_
 #define MUMMY_SHUFFLE_S4C_H_
 #define MUMMY_SHUFFLE_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for mummy_shuffle.
