+++ src/animations/srwarthog_square.c
