+++ src/animations/srwarthog_square.h
@@ -1,7 +1,7 @@
 #ifndef SRWARTHOG_SQUARE_S4C_H_
 #define SRWARTHOG_SQUARE_S4C_H_
 #define SRWARTHOG_SQUARE_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for srwarthog_square.
