+++ src/animations/troll_club.c
