+++ src/animations/troll_club.h
@@ -1,7 +1,7 @@
 #ifndef TROLL_CLUB_S4C_H_
 #define TROLL_CLUB_S4C_H_
 #define TROLL_CLUB_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for troll_club.
