+++ src/animations/werewolf_transform.c
