+++ src/animations/werewolf_transform.h
@@ -1,6 +1,7 @@
 #ifndef WEREWOLF_TRANSFORM_H_
+#define WEREWOLF_TRANSFORM_H_
 #define WEREWOLF_TRANSFORM_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for werewolf_transform.
