+++ src/animations/zombie_walk.c
