+++ src/animations/zombie_walk.h
@@ -1,7 +1,7 @@
 #ifndef ZOMBIE_WALK_S4C_H_
 #define ZOMBIE_WALK_S4C_H_
 #define ZOMBIE_WALK_S4C_H_VERSION "0.1.5"
-#include "../sprites4curses/s4c-animate/animate.h"
+#include "../../sprites4curses/s4c-animate/animate.h"
 
 /**
  * Declares animation matrix vector for zombie_walk.
