+++ src/game_curses.h
@@ -8,27 +8,27 @@
 #include "game_core.h"
 #include "game_utils.h"
 
-#include "knight_tapis.h"
-#include "mage_spark.h"
-#include "archer_drop.h"
-#include "assassin_poof.h"
+#include "./animations/knight_tapis.h"
+#include "./animations/mage_spark.h"
+#include "./animations/archer_drop.h"
+#include "./animations/assassin_poof.h"
 
-#include "mummy_shuffle.h"
-#include "ghost_spell.h"
-#include "boar_scream.h"
-#include "troll_club.h"
-#include "goblin_shoot.h"
-#include "zombie_walk.h"
-#include "imp_fireball.h"
-#include "werewolf_transform.h"
+#include "./animations/mummy_shuffle.h"
+#include "./animations/ghost_spell.h"
+#include "./animations/boar_scream.h"
+#include "./animations/troll_club.h"
+#include "./animations/goblin_shoot.h"
+#include "./animations/zombie_walk.h"
+#include "./animations/imp_fireball.h"
+#include "./animations/werewolf_transform.h"
 
-#include "crawlingdude_crawl.h"
-#include "srwarthog_square.h"
-#include "headlessninja_throw.h"
-#include "bluetroll_wonder.h"
+#include "./animations/crawlingdude_crawl.h"
+#include "./animations/srwarthog_square.h"
+#include "./animations/headlessninja_throw.h"
+#include "./animations/bluetroll_wonder.h"
 
-#include "enter_door.h"
-#include "alt_chest_opening.h"
+#include "./animations/enter_door.h"
+#include "./animations/alt_chest_opening.h"
 
 int display_colorpairs(void);
 void print_encoded_char(WINDOW* w, int y, int x, char c);
